+++ CHANGES.md
@@ -93,6 +93,7 @@
 
 * Support for X source added (Java/Python) ([#X](https://github.com/apache/beam/issues/X)).
 * MongoDB IO connector added (Go) ([#24575](https://github.com/apache/beam/issues/24575)).
+* Added in JmsIO a retry policy for failed publications (Java) ([#24971](https://github.com/apache/beam/issues/24971)).
 
 ## New Features / Improvements
 
@@ -136,7 +137,6 @@
 * S3 implementation of the Beam filesystem (Go) ([#23991](https://github.com/apache/beam/issues/23991)).
 * Support for SingleStoreDB source and sink added (Java) ([#22617](https://github.com/apache/beam/issues/22617)).
 * Added support for DefaultAzureCredential authentication in Azure Filesystem (Python) ([#24210](https://github.com/apache/beam/issues/24210)).
-* Added in JmsIO a retry policy for failed publications ([#24971](https://github.com/apache/beam/issues/24971)).
 
 ## New Features / Improvements